pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autolitdb"
version = "0.1.0"
description = "Automatic literature query, download, and analysis from diverse sources"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AutoLitDB Team"}
]
keywords = ["literature", "pubmed", "pdf", "rag", "llm", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    # HTTP & API
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    # Data processing
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
    # LLM clients
    "openai>=1.0.0",
    # RAG & Vector DB
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    # PDF processing
    "pymupdf>=1.23.0",
    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
    # Logging
    "loguru>=0.7.0",
    # Config
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    # Progress
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
autolitdb = "autolitdb.cli:main"

[project.urls]
Homepage = "https://github.com/autolitdb/autolitdb"
Repository = "https://github.com/autolitdb/autolitdb"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
